+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.10.1 (unreleased)
 
 - Added `if_not_exists` option to `create_namespace` method
+- Added `snapshot_id` option to `to_polars` method
 - Fixed installation error
 
 ## 0.10.0 (2025-08-27)